pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "3.11.9"
fastapi = "^0.112.0"
uvicorn = {extras = ["standart"], version = "^0.30.5"}
pydantic = {extras = ["email"], version = "^2.8.2"}
pydantic-settings = "^2.4.0"
prettytable = "^3.10.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
asyncpg = "^0.29.0"
alembic = "^1.13.2"
python-socketio = "^5.7.2"
python-jose = "^3.3.0"
passlib = "^1.7.4"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

